Payload path and client notifications (#170)
* wip adding alt payload path

* refactoring to add support for an alt payload path

* spliting out path for payload and thumbnails

* spliting out path for payload and thumbnails

* spliting out path for payload and thumbnails

* spliting out path for payload and thumbnails

* Adjusted notifications to encrypted when possible

* fixed test

* fixed test

---------

+++ services/Youverse.Core.Services/Drives/DriveCore/Storage/LongTermStorageManager.cs
@@ -10,20 +10,19 @@ using Youverse.Core.Services.Drives.FileSystem.Base;
 
 namespace Youverse.Core.Services.Drives.DriveCore.Storage
 {
-    public class FileBasedLongTermStorageManager : ILongTermStorageManager
+    public class LongTermStorageManager
     {
-        private readonly ILogger<ILongTermStorageManager> _logger;
+        private readonly ILogger<LongTermStorageManager> _logger;
 
         private readonly StorageDrive _drive;
 
         private const int WriteChunkSize = 1024;
 
-        // private const string ThumbnailDelimiter = "-";
         private const string ThumbnailDelimiter = "_";
         private const string ThumbnailSizeDelimiter = "x";
         private static readonly string ThumbnailSuffixFormatSpecifier = $"{ThumbnailDelimiter}{{0}}{ThumbnailSizeDelimiter}{{1}}";
 
-        public FileBasedLongTermStorageManager(StorageDrive drive, ILogger<ILongTermStorageManager> logger)
+        public LongTermStorageManager(StorageDrive drive, ILogger<LongTermStorageManager> logger)
         {
             Guard.Argument(drive, nameof(drive)).NotNull();
             // Guard.Argument(drive, nameof(drive)).Require(sd => Directory.Exists(sd.LongTermDataRootPath), sd => $"No directory for drive storage at {sd.LongTermDataRootPath}");
@@ -35,13 +34,23 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
             _drive = drive;
         }
 
+        /// <summary>
+        /// The drive managed by this instance
+        /// </summary>
         public StorageDrive Drive => _drive;
 
+        /// <summary>
+        /// Creates an Id for storing a file
+        /// </summary>
+        /// <returns></returns>
         public Guid CreateFileId()
         {
             return SequentialGuid.CreateGuid();
         }
 
+        /// <summary>
+        /// Writes a stream for a given file and part to the configured provider.
+        /// </summary>
         public Task WritePartStream(Guid fileId, FilePart part, Stream stream)
         {
             string filePath = GetFilenameAndPath(fileId, part, true);
@@ -49,13 +58,26 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
             return WriteFile(filePath, tempFilePath, stream);
         }
 
-        public Task DeleteFilePartStream(Guid fileId, FilePart filePart)
+        public Task DeletePayload(Guid fileId)
+        {
+            string path = GetFilenameAndPath(fileId, FilePart.Payload);
+            if(File.Exists(path))
+            {
+                File.Delete(path);
+            }
+            return Task.CompletedTask;
+        }
+
+        public Task DeleteFilePart(Guid fileId, FilePart filePart)
         {
             string path = GetFilenameAndPath(fileId, filePart);
             File.Delete(path);
             return Task.CompletedTask;
         }
 
+        /// <summary>
+        /// Gets a read stream for the given <see cref="FilePart"/>
+        /// </summary>
         public Task<Stream> GetFilePartStream(Guid fileId, FilePart filePart, FileChunk chunk = null)
         {
             string path = GetFilenameAndPath(fileId, filePart);
@@ -80,18 +102,23 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
                 {
                     Array.Resize(ref buffer, bytesRead);
                 }
-                
-                return Task.FromResult((Stream)new MemoryStream(buffer, false));
 
+                return Task.FromResult((Stream)new MemoryStream(buffer, false));
             }
 
             return Task.FromResult((Stream)fileStream);
         }
 
+        /// <summary>
+        /// Gets a read stream of the thumbnail
+        /// </summary>
+        /// <param name="fileId"></param>
+        /// <param name="width"></param>
+        /// <param name="height"></param>
         public Task<Stream> GetThumbnail(Guid fileId, int width, int height)
         {
             string fileName = GetThumbnailFileName(fileId, width, height);
-            string dir = GetFilePath(fileId, false);
+            string dir = GetFilePath(fileId, FilePart.Thumb);
             string path = Path.Combine(dir, fileName);
             var fileStream = File.Open(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
             return Task.FromResult((Stream)fileStream);
@@ -100,9 +127,9 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
         public Task DeleteThumbnail(Guid fileId, int width, int height)
         {
             string fileName = GetThumbnailFileName(fileId, width, height);
-            string dir = GetFilePath(fileId, false);
+            string dir = GetFilePath(fileId, FilePart.Thumb);
             string path = Path.Combine(dir, fileName);
-            
+
             File.Delete(path);
             return Task.CompletedTask;
         }
@@ -114,6 +141,10 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
             return fileName;
         }
 
+        /// <summary>
+        /// Ensures there is a valid file available for the given Id.
+        /// </summary>
+        /// <exception cref="InvalidDataException">Throw if the file for the given Id is invalid or does not exist</exception>
         public void AssertFileIsValid(Guid fileId)
         {
             if (fileId == Guid.Empty)
@@ -127,6 +158,9 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
             }
         }
 
+        /// <summary>
+        /// Checks if the file exists.  Returns true if all parts exist, otherwise false
+        /// </summary>
         public bool FileExists(Guid fileId)
         {
             return IsFileValid(fileId);
@@ -151,6 +185,9 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
             return true;
         }
 
+        /// <summary>
+        /// Removes all traces of a file and deletes its record from the index
+        /// </summary>
         public Task HardDelete(Guid fileId)
         {
             DeleteAllThumbnails(fileId);
@@ -165,6 +202,10 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
             return Task.CompletedTask;
         }
 
+        /// <summary>
+        /// Removes the contents of the meta file while permanently deletes the payload and thumbnails.  Retains some fields of the metafile and updates the index accordingly
+        /// </summary>
+        /// <param name="fileId"></param>
         public Task SoftDelete(Guid fileId)
         {
             DeleteAllThumbnails(fileId);
@@ -172,10 +213,14 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
             return Task.CompletedTask;
         }
 
-        public Task MoveToLongTerm(Guid targetFileId, string sourcePath, FilePart part)
+        /// <summary>
+        /// Moves the specified <param name="sourcePath"></param> to long term storage.
+        /// </summary>
+        public Task MovePayloadToLongTerm(Guid targetFileId, string sourcePath)
         {
+            FilePart part = FilePart.Payload;
             var dest = GetFilenameAndPath(targetFileId, part, ensureDirectoryExists: true);
-            Directory.CreateDirectory(Path.GetDirectoryName(dest));
+            Directory.CreateDirectory(Path.GetDirectoryName(dest) ?? throw new YouverseSystemException("Destination folder was null"));
 
             File.Move(sourcePath, dest, true);
             _logger.LogInformation($"File Moved to {dest}");
@@ -185,16 +230,21 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
         public Task MoveThumbnailToLongTerm(Guid targetFileId, string sourceThumbnail, int width, int height)
         {
             var dest = GetThumbnailPath(targetFileId, width, height);
-            Directory.CreateDirectory(Path.GetDirectoryName(dest));
+            Directory.CreateDirectory(Path.GetDirectoryName(dest) ?? throw new YouverseSystemException("Destination folder was null"));
             File.Move(sourceThumbnail, dest, true);
             _logger.LogInformation($"File Moved to {dest}");
 
             return Task.CompletedTask;
         }
 
+        /// <summary>
+        /// Returns an enumeration of <see cref="FileMetadata"/>; ordered by the most recently modified
+        /// </summary>
+        /// <param name="pageOptions"></param>
         public async Task<IEnumerable<ServerFileHeader>> GetServerFileHeaders(PageOptions pageOptions)
         {
-            string path = this.Drive.GetStoragePath(StorageDisposition.LongTerm);
+            // string path = this.Drive.GetStoragePath(StorageDisposition.LongTerm);
+            string path = this.Drive.GetLongTermHeaderStoragePath();
             var options = new EnumerationOptions()
             {
                 MatchCasing = MatchCasing.CaseInsensitive,
@@ -233,35 +283,34 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
             return header;
         }
 
-        public Task WriteThumbnail(Guid fileId, int width, int height, Stream stream)
-        {
-            var thumbnailPath = GetThumbnailPath(fileId, width, height);
-            var tempPath = GetTempFilePath(fileId, FilePart.Thumb, $"-{width}x{height}", true);
-
-            return WriteFile(thumbnailPath, tempPath, stream);
-        }
-
+        /// <summary>
+        /// Removes all thumbnails on disk which are not in the provided list.
+        /// </summary>
         public Task ReconcileThumbnailsOnDisk(Guid fileId, List<ImageDataHeader> thumbnailsToKeep)
         {
             Guard.Argument(thumbnailsToKeep, nameof(thumbnailsToKeep)).NotNull();
 
-            var thumbnailSearchPattern = string.Format(ThumbnailSuffixFormatSpecifier, "*", "*");
-            var seekPath = this.GetFilename(fileId, thumbnailSearchPattern, FilePart.Thumb);
-            string dir = GetFilePath(fileId, false);
+            string dir = GetFilePath(fileId, FilePart.Thumb);
 
-            var files = Directory.GetFiles(dir, seekPath);
-            foreach (var thumbnailFilePath in files)
+            if (Directory.Exists(dir))
             {
-                // filename w/o extension = "c1c63e18-40a2-9700-7b6a-2f1d51ee3972-300x300"
-                var filename = Path.GetFileNameWithoutExtension(thumbnailFilePath);
-                var sizeParts = filename.Split(ThumbnailDelimiter)[1].Split(ThumbnailSizeDelimiter);
-                var width = int.Parse(sizeParts[0]);
-                var height = int.Parse(sizeParts[1]);
-
-                var keepThumbnail = thumbnailsToKeep.Exists(thumb => thumb.PixelWidth == width && thumb.PixelHeight == height);
-                if (!keepThumbnail)
+                var thumbnailSearchPattern = string.Format(ThumbnailSuffixFormatSpecifier, "*", "*");
+                var seekPath = this.GetFilename(fileId, thumbnailSearchPattern, FilePart.Thumb);
+                
+                var files = Directory.GetFiles(dir, seekPath);
+                foreach (var thumbnailFilePath in files)
                 {
-                    File.Delete(thumbnailFilePath);
+                    // filename w/o extension = "c1c63e18-40a2-9700-7b6a-2f1d51ee3972-300x300"
+                    var filename = Path.GetFileNameWithoutExtension(thumbnailFilePath);
+                    var sizeParts = filename.Split(ThumbnailDelimiter)[1].Split(ThumbnailSizeDelimiter);
+                    var width = int.Parse(sizeParts[0]);
+                    var height = int.Parse(sizeParts[1]);
+
+                    var keepThumbnail = thumbnailsToKeep.Exists(thumb => thumb.PixelWidth == width && thumb.PixelHeight == height);
+                    if (!keepThumbnail)
+                    {
+                        File.Delete(thumbnailFilePath);
+                    }
                 }
             }
 
@@ -271,14 +320,14 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
         private string GetThumbnailPath(Guid fileId, int width, int height)
         {
             var thumbnailFileName = GetThumbnailFileName(fileId, width, height);
-            var filePath = GetFilePath(fileId);
+            var filePath = GetFilePath(fileId, FilePart.Thumb);
             var thumbnailPath = Path.Combine(filePath, thumbnailFileName);
             return thumbnailPath;
         }
 
-        private string GetFilePath(Guid fileId, bool ensureExists = false)
+        private string GetFilePath(Guid fileId, FilePart filePart, bool ensureExists = false)
         {
-            string path = _drive.GetStoragePath(StorageDisposition.LongTerm);
+            string path = filePart is FilePart.Payload or FilePart.Thumb ? _drive.GetLongTermPayloadStoragePath() : _drive.GetLongTermHeaderStoragePath();
 
             //07e5070f-173b-473b-ff03-ffec2aa1b7b8
             //The positions in the time guid are hex values as follows
@@ -310,13 +359,13 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
 
         private string GetFilenameAndPath(Guid fileId, FilePart part, bool ensureDirectoryExists = false)
         {
-            string dir = GetFilePath(fileId, ensureDirectoryExists);
+            string dir = GetFilePath(fileId, part, ensureDirectoryExists);
             return Path.Combine(dir, GetFilename(fileId, string.Empty, part));
         }
 
         private string GetTempFilePath(Guid fileId, FilePart part, string suffix, bool ensureExists = false)
         {
-            string dir = GetFilePath(fileId, ensureExists);
+            string dir = GetFilePath(fileId, part, ensureExists);
             string filename = $"{Guid.NewGuid()}{part}{suffix}.tmp";
             return Path.Combine(dir, filename);
         }
@@ -357,10 +406,10 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
         private void WriteStream(Stream stream, string filePath)
         {
             var buffer = new byte[WriteChunkSize];
-            var bytesRead = 0;
 
             using (var output = new FileStream(filePath, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.None))
             {
+                var bytesRead = 0;
                 do
                 {
                     bytesRead = stream.Read(buffer, 0, buffer.Length);
@@ -372,25 +421,19 @@ namespace Youverse.Core.Services.Drives.DriveCore.Storage
             }
         }
 
-        private void DeletePayload(Guid fileId)
-        {
-            string payload = GetFilenameAndPath(fileId, FilePart.Payload);
-            if (File.Exists(payload))
-            {
-                File.Delete(payload);
-            }
-        }
-
         private void DeleteAllThumbnails(Guid fileId)
         {
             var thumbnailSearchPattern = string.Format(ThumbnailSuffixFormatSpecifier, "*", "*");
             var seekPath = this.GetFilename(fileId, thumbnailSearchPattern, FilePart.Thumb);
-            string dir = GetFilePath(fileId, false);
+            string dir = GetFilePath(fileId, FilePart.Thumb);
 
-            var thumbnails = Directory.GetFiles(dir, seekPath);
-            foreach (var thumbnail in thumbnails)
+            if(Directory.Exists(dir))
             {
-                File.Delete(thumbnail);
+                var thumbnails = Directory.GetFiles(dir, seekPath);
+                foreach (var thumbnail in thumbnails)
+                {
+                    File.Delete(thumbnail);
+                }
             }
         }
     }
